Changed useconds_t and suseconds_t to be "long" types instead of "long long". This has a negative effect on ualarm(), but is somewhat required by the POSIX specs ("[...] the implementation should support one or more programming environments in which [...], and useconds_t are no greater than the width of type long." Indeed, it also changed the size of our struct timeval which broken binary compatibility with R5 (ie. used in gettimeofday()). This fixes bug #490 and therefore bug #428 (firefox crashing on load) as well.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@17142 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/posix/time.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2005, Haiku Inc. All Rights Reserved.
+ * Copyright 2005-2006, Haiku Inc. All Rights Reserved.
  * Distributed under the terms of the MIT License.
  */
 #ifndef _TIME_H_
@@ -11,8 +11,8 @@
 
 typedef long clock_t;
 typedef long time_t;
-typedef long long suseconds_t;
-typedef unsigned long long useconds_t;
+typedef long suseconds_t;
+typedef unsigned long useconds_t;
 
 #define CLOCKS_PER_SEC	1000
 #define CLK_TCK			CLOCKS_PER_SEC
